Fix transfer schedule cache update logic

+++ pe-common/nfs-near-cache/core-near-cache-consumer/src/main/java/com/nextuple/core/consumer/LocalCacheUpdateService.java
@@ -11,7 +11,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import lombok.RequiredArgsConstructor;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -54,28 +57,72 @@ public class LocalCacheUpdateService {
           Constructor<?> cons = c.getConstructor();
           var cacheKey = (CacheKey) cons.newInstance();
 
-          String path = "nearcache.entity." + entity + ".attributes";
-          String params = env.getProperty(path);
+          if (className.equals("com.nextuple.transit.cache.domain.TransferScheduleCacheKey")) {
+            createCacheKeyForTransferSchedules(
+                genericNearCacheService, message, className, cacheKey);
+          } else {
+            String path = "nearcache.entity." + entity + ".attributes";
+            String params = env.getProperty(path);
 
-          logger.debug("Params list :{}", params);
+            logger.debug("Params list :{}", params);
 
-          List<String> paramsList = new ArrayList<>();
-          if (!ObjectUtils.isEmpty(params)) {
-            paramsList = Arrays.asList(params.split("\\s*,\\s*")); // NOSONAR
+            List<String> paramsList = new ArrayList<>();
+            if (!ObjectUtils.isEmpty(params)) {
+              paramsList = Arrays.asList(params.split("\\s*,\\s*")); // NOSONAR
+            }
+            for (String param : paramsList) {
+              var field = c.getDeclaredField(param);
+              castToRequiredType(message, param, field);
+              field.setAccessible(true); // NOSONAR
+              field.set(cacheKey, message.get(param)); // NOSONAR
+            }
+            logger.debug("Cache key :{}", cacheKey);
+
+            genericNearCacheService.delete(cacheKey); // NOSONAR
           }
-          for (String param : paramsList) {
-            var field = c.getDeclaredField(param);
-            castToRequiredType(message, param, field);
-            field.setAccessible(true); // NOSONAR
-            field.set(cacheKey, message.get(param)); // NOSONAR
-          }
-
-          logger.debug("Cache key :{}", cacheKey);
-
-          genericNearCacheService.delete(cacheKey); // NOSONAR
         }
       }
     }
+  }
+
+  private void createCacheKeyForTransferSchedules(
+      GenericNearCacheService genericNearCacheService,
+      Map<String, Object> message,
+      String className,
+      CacheKey cacheKey)
+      throws ClassNotFoundException, IllegalAccessException, NoSuchFieldException {
+    Class<?> c = Class.forName(className);
+
+    Map<String, String> paramsToMessageFieldMap =
+        Map.of(
+            "orgId",
+            "orgId",
+            "rule",
+            "rule",
+            "ruleName",
+            "ruleName",
+            "dropoffNode",
+            "dropoffNodeId");
+
+    for (Entry<String, String> paramEntry : paramsToMessageFieldMap.entrySet()) {
+      var field = c.getDeclaredField(paramEntry.getKey());
+      castToRequiredType(message, paramEntry.getValue(), field);
+      field.setAccessible(true); // NOSONAR
+      field.set(cacheKey, message.get(paramEntry.getValue())); // NOSONAR
+    }
+
+    // set date bucket here
+    DateTime startDateTime = new DateTime(message.get("startTime")).toDateTime(DateTimeZone.UTC);
+    DateTime endDateTime = new DateTime(message.get("endTime")).toDateTime(DateTimeZone.UTC);
+
+    var field = c.getDeclaredField("dateBucket");
+    field.setAccessible(true);
+    field.set(cacheKey, startDateTime.withTime(0, 0, 0, 0).toDate());
+
+    genericNearCacheService.delete(cacheKey); // NOSONAR
+
+    field.set(cacheKey, endDateTime.withTime(23, 59, 59, 0).toDate());
+    genericNearCacheService.delete(cacheKey); // NOSONAR
   }
 
   private static void castToRequiredType(Map<String, Object> message, String param, Field field) {
